# Reduce memory allocations in network layer (3d edition)