--- a/FO-DICOM.Core/IO/Buffer/RangeByteBuffer.cs
+++ b/FO-DICOM.Core/IO/Buffer/RangeByteBuffer.cs
@@ -1,8 +1,9 @@
 ﻿// Copyright (c) 2012-2021 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 
+using FellowOakDicom.Memory;
+using Microsoft.Extensions.DependencyInjection;
 using System;
-using System.Buffers;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
@@ -11,8 +12,16 @@ namespace FellowOakDicom.IO.Buffer
 {
     public class RangeByteBuffer : IByteBuffer
     {
-        public RangeByteBuffer(IByteBuffer buffer, long offset, int length)
+        private readonly IMemoryProvider _memoryProvider;
+
+        public RangeByteBuffer(IByteBuffer buffer, long offset, int length): this(buffer, offset, length, Setup.ServiceProvider.GetRequiredService<IMemoryProvider>())
         {
+            
+        }
+        
+        public RangeByteBuffer(IByteBuffer buffer, long offset, int length, IMemoryProvider memoryProvider)
+        {
+            _memoryProvider = memoryProvider ?? throw new ArgumentNullException(nameof(memoryProvider));
             Internal = buffer;
             Offset = offset;
             Length = length;
@@ -54,25 +63,18 @@ namespace FellowOakDicom.IO.Buffer
             }
 
             int bufferSize = 1024 * 1024;
-            byte[] buffer = ArrayPool<byte>.Shared.Rent(bufferSize);
+            using IMemory buffer = _memoryProvider.Provide(bufferSize);
             long remaining = Size;
             long offset = 0;
-            try
+            while (offset < remaining)
             {
-                while (offset < remaining)
-                {
-                    var count = (int)Math.Min(remaining - offset, bufferSize);
+                var count = (int)Math.Min(remaining - offset, bufferSize);
 
-                    GetByteRange(offset, count, buffer);
+                GetByteRange(offset, count, buffer.Bytes);
 
-                    stream.Write(buffer, 0, count);
+                stream.Write(buffer.Bytes, 0, count);
 
-                    offset += count;
-                }
-            }
-            finally
-            {
-                ArrayPool<byte>.Shared.Return(buffer);
+                offset += count;
             }
         }
 
@@ -89,25 +91,18 @@ namespace FellowOakDicom.IO.Buffer
             }
 
             int bufferSize = 1024 * 1024;
-            byte[] buffer = ArrayPool<byte>.Shared.Rent(bufferSize);
+            using IMemory buffer = _memoryProvider.Provide(bufferSize);
             long size = Size;
             long offset = 0;
-            try
+            while (offset < size)
             {
-                while (offset < size)
-                {
-                    var count = (int)Math.Min(size - offset, bufferSize);
+                var count = (int)Math.Min(size - offset, bufferSize);
 
-                    GetByteRange(offset, count, buffer);
+                GetByteRange(offset, count, buffer.Bytes);
 
-                    await stream.WriteAsync(buffer, 0, count, cancellationToken);
+                await stream.WriteAsync(buffer.Bytes, 0, count, cancellationToken);
 
-                    offset += count;
-                }
-            }
-            finally
-            {
-                ArrayPool<byte>.Shared.Return(buffer);
+                offset += count;
             }
         }
     }