--- a/FO-DICOM.Core/Memory/ArrayPoolMemory.cs
+++ b/FO-DICOM.Core/Memory/ArrayPoolMemory.cs
@@ -1,4 +1,7 @@
 ﻿using Microsoft.Toolkit.HighPerformance.Buffers;
+// Copyright (c) 2012-2021 fo-dicom contributors.
+// Licensed under the Microsoft Public License (MS-PL).
+
 using System;
 
 namespace FellowOakDicom.Memory