--- a/FO-DICOM.Core/Memory/ArrayPoolMemoryProvider.cs
+++ b/FO-DICOM.Core/Memory/ArrayPoolMemoryProvider.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.Toolkit.HighPerformance.Buffers;
+
+namespace FellowOakDicom.Memory
+{
+    public class ArrayPoolMemoryProvider : IMemoryProvider
+    {
+        public IMemory Provide(int length) => new ArrayPoolMemory(MemoryOwner<byte>.Allocate(length));
+    }
+}