--- a/FO-DICOM.Core/Memory/IMemory.cs
+++ b/FO-DICOM.Core/Memory/IMemory.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) 2012-2021 fo-dicom contributors.
+// Licensed under the Microsoft Public License (MS-PL).
+
+using System;
 
 namespace FellowOakDicom.Memory
 {