The next batch of performance / memory improvements. This will focus on DicomService and PDataTFStream.

1st edition: #1267
2nd edition: #1273

Related issue: #1262 

#### Checklist
- [x] The pull request branch is in sync with latest commit on the *fo-dicom/development* branch
- [x] I have updated API documentation
- [x] I have included unit tests
- [x] I have updated the change log
- [x] I am listed in the CONTRIBUTORS file

#### Changes proposed in this pull request:
- Add dependency on Microsoft.Toolkit.HighPerformance.Buffers and use MemoryOwner
- Use rented byte arrays when buffering incoming raw PDUs (implementation detail: only when PDU < 1 MB)
- Avoid intermediate buffer copy in DicomService
- Use rented byte arrays inside PDataTFStream when writing P-DATA-TF
- Make separate classes for dependencies of DicomServer and DicomService so we can change these dependencies later without breaking consumers
- Introduce new abstraction: IMemoryProvider and replace all usages of ArrayPool<byte>.Shared with this new IMemoryProvider