@gofal 2021-12-16
I see that we are now using ArrayPool, because of reusing rented arrays and therefore reducing memory allocations. But this now changes code pretty much, because we have to take care about returning the rented array in finalizers, and the rented array may be larger then the requested length and therefore we have to store the requested length in separate fields.
When reviewing the code, I thought, this must be simpler, by wrapping the rented array into a IDisposable class. And actually, Microsoft did that already for us:
https://docs.microsoft.com/en-us/windows/communitytoolkit/high-performance/memoryowner
Have you already taken a look at that class? using the MemoryOwner<T> instead of rented arrays may make have the same memory improvement, but make the code better readable again - and since the rented memory is wrapped, it is more save to avoid memory leaks in case we forget to release the rented array somewhere.