@amoerie 2021-12-16
I agree that bare ArrayPool usage is a little "low level", and it has pitfalls, but for cases where the array is only used locally and is very short lived (like in most usages inside the PDU file) I think it is fine as the entire lifetime of the array is contained inside a try { ... } finally { ... } block where the rented array is returned inside the finally block.

However, if the lifetime of the array becomes more vague, I agree that some sort of abstraction would be very useful so we can implement safety nets using finalizers. I haven't used MemoryOwner yet, but I agree that it looks like an ideal solution for how we're using ArrayPool. Is it okay to add this new dependency then? Looks like it supports .NET Standard 2.0, so we definitely could.

Furthermore, I wanted  to ask you whether it might be a good idea to make a "MemoryManager", much in the same way that we already have a NetworkManager, LogManager, etc. today. This would allow expert consumers of fo-dicom to swap in their own memory allocation implementation. ( Perhaps something based on https://github.com/microsoft/Microsoft.IO.RecyclableMemoryStream ) 
We could provide a default implementation using MemoryOwner from the high performance toolkit. 

I imagine the interface would look a little like this:

```csharp
public interface IMemoryManager {
  IMemory GetMemory(int minimumLength);
}

public interface IMemory : IDisposable {
  byte[] Bytes { get; }
  int MinimumLength { get; }
}
```
(Note: Microsoft is pushing towards using Span<byte> and Memory<byte> instead of byte[], but I'm not confident all of our usages would be compatible)

To be clear, I would not always use this memory manager. ArrayPool<byte>.Shared is highly optimized and safe to always use. 
However, its design is optimized for small (< 1MB), short-lived arrays. Short-lived means things like intermediate buffers mostly. 

In other words, I would still prefer to use ArrayPool<byte>.Shared in cases where the buffer fits these criteria perfectly. Examples for fo-dicom would include things like PDU where we write an association request.

However, the MemoryManager would be preferred if the lifetime can get more complex and the size can get bigger. Examples for fo-dicom would include things like PDataTFStream and incoming Raw PDUs in DicomService.
Further down the road, we could look at image rendering and other places too, but that might require public signature changes.