@gofal 2021-12-16
A Reference to Microsofts HighPerformance package is fine. This package is supported from netstandard1.4 upwards. 

Sure, when renting a 6 byte array for the preamble, then the overhead of a wrapping MemoryOwner would be too much.

But I even would prefer using the MemoryOwner, or some similar wrapper over the rented arrays from ArrayPools in cases where both would fit and the overhead of the MemoryOwner is not breaking performance. Because the simple fact that the rented array may be larger than the requested size and you have to store somewhere the actual length is worth the wrapper and helps preventing future bugs.

Having the MemoryManager would be fine, if we really everywhere would use an Interface instead of byte arrays. Eg if we would Streams always instead of bytes arrays. Or Span, which will be available once we drop support for netframework. 

The Manager-classes are kind of deprecated and should be replaced by the DI-Container. So there is absolute nothing to say against registering the MemoryStream as IMemoryStream and instead of creating a new instance of MemoryStream, requesting a service from the serviceprovider. Then a user can register his own implementation of IMemoryStream, like the RecycalbeMemoryStream.