@gofal 2021-12-16
You are now optimizing Memory in netwok communication. 
Optimizing memory when it comes to loading/storing, there are many temporary streams, most of them MemoryStreams, that internally always allocate new byte arrays. Or when Decoding is done, then many parts of the memory are copied. So there will be some place for optimization. 
I do have an issue, that writing performance of WSI files (some gigabytes) is terrible, because everything is kept in memory. i will have to do quite some optimizations there, but i don't want to have code conflicts, thats why i am waiting for you to be done with that all, before I start.