@gofal 2021-12-16
concerning the MemoryManager or MemoryProvider. The naming does not mind, but the huge difference. the manager is one static thing, while with DI you can configure and inject more precisely. If you are running within a console application, then the DI also breaks down into one static Setup.ServiceProvider. But as soon as you are running in asp.net or some other environment with built in DI, this gives you far more control and possiblities.

The various services, that the user can override via DI, are documented in read.me (this are only very few at the moment), and of course there are already extension methods for those.

But to really give the user the possibility to inject his own memory implementation, we would have to use it very extensively. Namely at least everywhere, where you now used ArrayPool, and everywhere, where a new MemoryStream is instanciated etc. And the interface requires quite some entry-points, like allocating, freeing, maybe locking, etc. I doubt if it's worth it, at least not in such core code like here, where you are doing optimizations anyway.

Maybe when it comes to long-lasting memory, like when reading a file, where we now use MemoryStream. Maybe someone wants to inject there and use TempFileStream for all of those to reduce the heap size dramatically.