@gofal 2021-12-16
> Furthermore, each DicomItem in a DicomDataset gets his own IByteBuffer allocated. It might be a lot more efficient to have one ReadonlySequence of byte to represent the Dicom file, and then give every DicomItem a ReadonlyMemory of byte that represents its contents.
However, all of this would only work if we introduce the concept of a "ReadonlyDicomDataset" which could be a lot more performant and simpler because it does not have to be writable.

some time ago, i managed to reduce memory drastically, by extending the IBytebuffer with some stream-methods. Previously, there have always been new byte arrays where data is copied many times. with the streams, now one MemoryStream is holding the actual byte-array, and on top of those there are streams only pointing on parts of this memorystream. A similar concpet as there is now with Span. But Span is more low-level, since it can also work with pinned unmanaged memory. But having streams more consequently will already improve a lot.