@amoerie 2022-02-08
For sport, these are benchmarks for sending an 11MB DICOM file (Mammo)

**Before**
-----------


``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.22000
Intel Core i7-8700K CPU 3.70GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=6.0.101
  [Host]     : .NET Core 3.1.22 (CoreCLR 4.700.21.56803, CoreFX 4.700.21.57101), X64 RyuJIT
  Job-VEQQYQ : .NET Core 3.1.22 (CoreCLR 4.700.21.56803, CoreFX 4.700.21.57101), X64 RyuJIT

InvocationCount=128  MaxIterationCount=25  MaxWarmupIterationCount=10  
UnrollFactor=16  

```
|            Method |      Mean |     Error |    StdDev |     Gen 0 |     Gen 1 |     Gen 2 |   Allocated |
|------------------ |----------:|----------:|----------:|----------:|----------:|----------:|------------:|
| SendStoreToServer | 21.949 ms | 1.3789 ms | 1.8408 ms | 1117.1875 | 1070.3125 | 1062.5000 | 46918.83 KB |


**After**
-----------

``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.22000
Intel Core i7-8700K CPU 3.70GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=6.0.101
  [Host]     : .NET Core 3.1.22 (CoreCLR 4.700.21.56803, CoreFX 4.700.21.57101), X64 RyuJIT
  Job-XNQFXO : .NET Core 3.1.22 (CoreCLR 4.700.21.56803, CoreFX 4.700.21.57101), X64 RyuJIT

InvocationCount=128  MaxIterationCount=25  MaxWarmupIterationCount=10  
UnrollFactor=16  

```
|            Method |      Mean |     Error |    StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------ |----------:|----------:|----------:|--------:|------:|------:|----------:|
| SendStoreToServer | 20.377 ms | 2.9066 ms | 3.8802 ms | 46.8750 |     - |     - | 321.19 KB |

So yay, we don't even allocate 1MB! We finally have full streaming of a DICOM file from client to server without ever loading the full DICOM file in memory. 🎉